Change shell prompts to read "python3" vs "python"
Since this is a tutorial written largely for novices, it is more user friendly to have the shell prompts be "python3". The tutorial states that "This tutorial is written for Django |version|, which supports Python 3.8 and later." So it is safe to assume the user is using python3 for shell prompts. Any experienced python developer who is using an alias such as python-is-python3 will now to substitute python for python3, but a novice will not necessarily know to do the inverse.